Use safe matcher while pulling test suites
Use a safe matcher while pulling test suites.
If a string fails to evaluate in the given
duration, it is thrown out. If all test suite
strings are thrown out, the whole test suite
is dropped. The reasoning here is that if a
regex can't evaluate any of its strings, then
the test suite is insane