Update evaluation system to do partial & full match at the same time
Here's now the evaluation works now:
- don't filter in `EvaluationService`. Move
that down to the test suite evaluator
- there are now two methods: satisfiesFull and satisfiesPartial. The
logics are a little bit different, but basically full just checks if
full string matches, while partial actually checks to see if the
correct substring is pulled out
- if a test suite has both full +/- strings, then satisfiesFull is
carried out. If the test suite does not satisfy the filter, then an
indeterminate boolean is used, i.e., "i don't know cause i couldn't
check"
- Same deal as above for partial
- Each result has an indeterminate "full" and "partial" match. Results
are filtered by if they have a true value for either cell. If both are
either indeterminate or false, the result is ignored
- These results are then saved like usual