Update DistanceScore caching strategy
A few changes:
1. split tree cache into two parts: truth and candidate. Candidate is larger
and holds all regexes, while truth is smaller. The idea is that each record
will be grouped by truth regex, so we can keep a smaller cache for that
one
2. added a score cache. If we see the same pair of (truth, candidate)
regexes, then we can use a memoized score to skip building trees and
recomputing distance. THIS ASSUMES THAT THE TWO REGEXES ARE ALWAYS IDENTICAL,
WHICH IS A RISK I'M WILLING TO TAKE FOR NOW

TODO: i should really revisit how i wrote that stream. It's a little
goofy